Rev: Refactored the lasers in the PuzzleSystem
- The lasers in the puzzle system are not projectiles that can be used by characters

+++ Assets/_Main/Scripts/GamePlay/AmmunationSystem/AmmoCounter.cs
@@ -47,7 +47,7 @@ namespace _Main.Scripts.GamePlay.AmmunitionSystem
         }
 
         #region IVisualizable
-        public Action<int> OnValueChanged { get; set; }
+        public Action<float> OnValueChanged { get; set; }
         public Action OnMaxValueChanged { get; set; }
         public Action OnClose { get; set; }
