+++ Assets/_Main/Scripts/GamePlay/UI/IVisualizable.cs
@@ -5,7 +5,7 @@ namespace _Main.Scripts.GamePlay.UI
 {
     public interface IVisualizable
     {
-        public Action<int> OnValueChanged { get; set; }
+        public Action<float> OnValueChanged { get; set; }
         public Action OnMaxValueChanged { get; set; }
         public Action OnClose { get; set; }
